pyproject.toml
[project]
name = "adk-auth-exp"
version = "0.1.0"
description = "ADK Agent with Spanner Toolkit - Auth Experiment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.24.0",
    "google-cloud-spanner>=3.0.0",
]

[[tool.uv.index]]
url = "https://pypi.org/simple/"
default = true
